+++ Private/Core/Save-TheaterState.ps1
@@ -5,7 +5,7 @@ function Save-TheaterState {
     [CmdletBinding()]
     param(
         [Parameter(Mandatory)]
-        [ValidateSet('entra', 'ad', 'm365')]
+        [ValidateSet('entra', 'ad', 'm365', 'workspace')]
         [string]$Theater,
 
         [Parameter(Mandatory)]